+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Updating the history of `wal-project-*` commands now uses fuzzy
+  matching instead of just checking the last `compile-command` (as
+  this can lead to erroneous replacements).
+
 ## [v2.2.1]
 
 Itchy and patchy.